Closing the window without using the Exit button now triggers a save check.
